+++ packages/code-studio/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.1](https://github.com/deephaven/web-client-ui/compare/v1.18.0...v1.18.1) (2026-04-30)
+
+**Note:** Version bump only for package @deephaven/code-studio
+
+## [1.18.0](https://github.com/deephaven/web-client-ui/compare/v1.17.1...v1.18.0) (2026-04-09)
+
+### Bug Fixes
+
+- DH-21539: Prevent browser back navigation on horizontal overscroll ([#2654](https://github.com/deephaven/web-client-ui/issues/2654)) ([d504ec8](https://github.com/deephaven/web-client-ui/commit/d504ec853c6434343ea544b0ccee7710e666881b))
+
 ## [1.17.1](https://github.com/deephaven/web-client-ui/compare/v1.17.0...v1.17.1) (2026-04-01)
 
 **Note:** Version bump only for package @deephaven/code-studio